refactor: organised routes and gives status code with error msg
