+++ .env
@@ -1,3 +1,1 @@
-REDIS_URI=redis://redistogo:f99edf3de0cyryty324fe462@sole.redistogo.com:10393/
-PYTHON_ENV=prod
-CACHE_EXPIRATION=180
+#!ENV NOT NEEDED
